requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.1
PyQt6==6.4.0
PyQt6-Charts==6.4.0
PyQt6-Charts-Qt6==6.4.2
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.0
python-dotenv==0.21.0
